Add TruncateHtmlPipe for text length limit
Introduced a new pipe in Angular, named 'truncateHtml', which can limit the length of the displayed text. This pipe removes all HTML tags from the text and trims it to the specified number of characters. It will be useful for displaying shortened versions of potentially long strings of text. Unit tests for this new pipe have also been added.